swapped index&h-r
